# prov/shm: always call SAR progress function