SAR can be used before the CMA check occurs and is also used for device
transfers. Call this function regardless whether CMA is available.

Signed-off-by: Robert Wespetal <wesper@amazon.com>