@rwespetal 2020-09-29
Here's a bit more information about the test that I ran, hopefully this helps with reproducing on your side. These hosts have 36 cores, so rank 37 is the only rank on the second node. ptrace protection doesn't affect the test result here.

With this patch:
```
$ I_MPI_DEBUG=1 mpirun -n 37 -ppn 36 -hostlist 172.31.8.98,172.31.14.236 -prepend-rank ./mpi/collective/osu_alltoall
[0] [0] MPI startup(): libfabric version: 1.12.0a1
[0] [0] MPI startup(): libfabric provider: efa
[0] 
[0] # OSU MPI All-to-All Personalized Exchange Latency Test v5.6.3
[0] # Size       Avg Latency(us)
[0] 1                     131.90
[0] 2                     134.65
[0] 4                     136.12
[0] 8                     223.33
[0] 16                    162.02
[0] 32                    178.55
[0] 64                    262.17
[0] 128                    84.80
[0] 256                   703.79
[0] 512                   146.55
[0] 1024                   67.97
[0] 2048                   77.41
[0] 4096                   95.15
[0] 8192                  148.12
[0] 16384                 312.62
[0] 32768                 992.99
[0] 65536                2508.01
[0] 131072               5117.71
[0] 262144              10354.86
[0] 524288              17850.11
[0] 1048576             38247.82
```
Without this patch:
```
$ I_MPI_DEBUG=1 mpirun -n 37 -ppn 36 -hostlist 172.31.8.98,172.31.14.236 -prepend-rank ./mpi/collective/osu_alltoall
[0] [0] MPI startup(): libfabric version: 1.12.0a1
[0] [0] MPI startup(): libfabric provider: efa
[0] 
[0] # OSU MPI All-to-All Personalized Exchange Latency Test v5.6.3
[0] # Size       Avg Latency(us)
[0] 1                     131.90
[0] 2                     131.58
[0] 4                     133.58
[0] 8                     147.51
[0] 16                    161.40
[ ... hang ... ]
```

And adding a print in the SAR case without this patch showing the cma_cap is not set:
```
$ I_MPI_DEBUG=1 mpirun -n 37 -ppn 36 -hostlist 172.31.8.98,172.31.14.236 -prepend-rank ./mpi/collective/osu_alltoall -i 1 -x 0 -m 32:32
[0] [0] MPI startup(): libfabric version: 1.12.0a1[0]
[0] [0] MPI startup(): libfabric provider: efa[0]
[0] 
[0] # OSU MPI All-to-All Personalized Exchange Latency Test v5.6.3
[0] # Size    [0]    Avg Latency(us)[0]
[36] == smr_generic_rma using sar len 41472 cma_cap 0
[ ... hang ... ]
```

```
$ mpirun --version
Intel(R) MPI Library for Linux* OS, Version 2019 Update 7 Build 20200312 (id: 5dc2dd3e9)
Copyright 2003-2020, Intel Corporation.
```